begin specs for ingredient manager service and migrate summy data from main menu
