+++ libs/apps/ingredient-manager/src/lib/ingredient-manager/views/main-menu/main-menu.component.ts
@@ -31,93 +31,6 @@ USER STORY:
 
 
 
-const DemoCategory : CategoryIndex[] = [
-  {
-    name: 'Salts',
-    id: 'salts_category_id1100011',
-    subCategories: [
-      {
-        name: 'Sea Salts',
-        id: 'sea_salts_id220022',
-        items: [
-          {
-            name: 'sea salt 01',
-            id: 'sea_salt_01',
-            icon: 'some-icon'
-          },
-          {
-            name: 'sea salt 02',
-            id: 'sea_salt_02',
-            photo: 'some-photo'
-          },
-          {
-            name: 'sea salt 03',
-            id: 'sea_salt_03',
-            icon: 'some-icon'
-          },
-          {
-            name: 'sea salt 04',
-            id: 'sea_salt_04',
-            icon: 'some-icon'
-          },
-          {
-            name: 'sea salt 05',
-            id: 'sea_salt_05',
-            photo: 'some-photo'
-          },
-          {
-            name: 'sea salt 06',
-            id: 'sea_salt_06',
-            icon: 'some-icon'
-          },
-        ],
-        icon: 'some-icon'
-      },
-      {
-        name: 'Kosher Salts',
-        id: 'kosher_salts_id220022',
-        items: [
-          {
-            name: 'kosher salt 01',
-            id: 'kosher_salt_01',
-            icon: 'some-icon'
-          },
-          {
-            name: 'kosher salt 02',
-            id: 'kosher_salt_02',
-            photo: 'some-photo'
-          }
-        ],
-        icon: 'some-icon'
-      },
-      {
-        name: 'Table Salts',
-        id: 'table_salts_id220022',
-        items: [
-          {
-            name: 'table salt 01',
-            id: 'table_salt_01',
-            icon: 'some-icon'
-          },
-          {
-            name: 'table salt 02',
-            id: 'table_salt_02',
-            photo: 'some-photo'
-          },
-          {
-            name: 'table salt 03',
-            id: 'table_salt_03',
-            photo: 'some-photo'
-          }
-        ],
-        icon: 'some-icon'
-      }
-    ]
-  }
-]
-
-
-
 
 
 import { Component } from '@angular/core';
